Move <Content/> components into shared layout folder
