+++ Source/SelfService/Web/integrations/connection/connectionDetails/messages/changeMessage/index.tsx
@@ -12,7 +12,7 @@ import {
 import { useConnectionId } from '../../../../routes.hooks';
 
 import { ViewMode } from './ViewMode';
-import { ContentContainer } from './components/ContentContainer';
+import { ContentContainer } from '../../../../../components/layout/Content/ContentContainer';
 import { MessageMappingForm } from './components/MessageMappingForm';
 import { ChangeMessageView } from './ChangeMessageView';
